Refactoring: simplify game object. Add user authorization.

+++ api/src/routes/Games.ts
@@ -1,11 +1,11 @@
 import { Request, Response, Router } from 'express';
-import { BAD_REQUEST, CREATED, OK, NOT_FOUND } from 'http-status-codes';
+import { BAD_REQUEST, CREATED, OK, NOT_FOUND, FORBIDDEN } from 'http-status-codes';
 import { ParamsDictionary } from 'express-serve-static-core';
 import { generateId } from '@shared/functions';
 
 import GameDao from '@daos/Game/GameDao.mock';
 import * as GameController from '@entities/Game';
-import { paramMissingError, gameNotFoundError } from '@shared/constants';
+import { paramMissingError, gameNotFoundError, forebiddenError } from '@shared/constants';
 
 // Init shared
 const router = Router();
@@ -17,7 +17,11 @@ const gameDao = new GameDao();
  ******************************************************************************/
 
 router.get('/all', async (req: Request, res: Response) => {
-    const games = await gameDao.getAll();
+    const userId = req.headers.authorization;
+    let games = await gameDao.getAll();
+    games = games.filter((game: GameController.IGame) => {
+        return game.phase === GameController.GamePhase.Init || (userId && game.players.findIndex(p => p.id === userId) > -1);
+    });
     return res.status(OK).json({games});
 });
 
@@ -36,6 +40,7 @@ router.get('/:id', async (req: Request, res: Response) => {
 
 router.post('/add', async (req: Request, res: Response) => {
     const { game } = req.body;
+    const userId = req.headers.authorization;
     if (!game) {
         return res.status(BAD_REQUEST).json({
             error: paramMissingError,
@@ -43,7 +48,7 @@ router.post('/add', async (req: Request, res: Response) => {
     }
 
     if (!game.id) game.id = generateId();
-    if (!game.host) game.host = generateId();
+    if (!game.host) game.host = userId;
 
     await gameDao.add(game);
     return res.status(CREATED).json({id: game.id, playerId: game.host});
@@ -71,10 +76,16 @@ router.put('/update', async (req: Request, res: Response) => {
 
 router.put('/:id/startPreparation', async (req: Request, res: Response) => {
     const { wordsPerPlayer } = req.body;
+    const userId = req.headers.authorization;
     const game = await gameDao.getOne(req.params.id);
     if (!game) {
         return res.status(NOT_FOUND).json({
             error: gameNotFoundError,
+        });
+    }
+    if (game.host !== userId) {
+        return res.status(FORBIDDEN).json({
+            error: forebiddenError,
         });
     }
 
@@ -90,6 +101,7 @@ router.put('/:id/startPreparation', async (req: Request, res: Response) => {
 
 router.put('/:id/addPlayer', async (req: Request, res: Response) => {
     const { player } = req.body;
+    const userId = req.headers.authorization;
     const game = await gameDao.getOne(req.params.id);
     if (!player) {
         return res.status(BAD_REQUEST).json({
@@ -102,7 +114,7 @@ router.put('/:id/addPlayer', async (req: Request, res: Response) => {
         });
     }
 
-    if (!player.id) player.id = generateId();
+    if (!player.id) player.id = userId;
     GameController.addPlayer(game, player);
 
     await gameDao.update(game);
@@ -115,10 +127,16 @@ router.put('/:id/addPlayer', async (req: Request, res: Response) => {
 
 router.put('/:id/updatePlayer', async (req: Request, res: Response) => {
     const { player } = req.body;
+    const userId = req.headers.authorization;
     const game = await gameDao.getOne(req.params.id);
     if (!player) {
         return res.status(BAD_REQUEST).json({
             error: paramMissingError,
+        });
+    }
+    if (player.id !== userId) {
+        return res.status(FORBIDDEN).json({
+            error: forebiddenError,
         });
     }
     if (!game) {
@@ -139,6 +157,7 @@ router.put('/:id/updatePlayer', async (req: Request, res: Response) => {
 
 router.put('/:id/hint', async (req: Request, res: Response) => {
     const { hint } = req.body;
+    const userId = req.headers.authorization;
     const game = await gameDao.getOne(req.params.id);
     if (!hint) {
         return res.status(BAD_REQUEST).json({
@@ -150,8 +169,13 @@ router.put('/:id/hint', async (req: Request, res: Response) => {
             error: gameNotFoundError,
         });
     }
+    if (game.players.findIndex(p => p.id === userId) === -1) {
+        return res.status(FORBIDDEN).json({
+            error: forebiddenError,
+        });
+    }
 
-    GameController.addHint(game, hint.hint, hint.author.id);
+    GameController.addHint(game, hint, userId);
 
     await gameDao.update(game);
     return res.status(OK).end();
@@ -163,10 +187,16 @@ router.put('/:id/hint', async (req: Request, res: Response) => {
 
 router.put('/:id/toggleDuplicateHint', async (req: Request, res: Response) => {
     const { hintIndex } = req.body;
+    const userId = req.headers.authorization;
     const game = await gameDao.getOne(req.params.id);
     if (!game) {
         return res.status(NOT_FOUND).json({
             error: gameNotFoundError,
+        });
+    }
+    if (game.roundHost !== userId) {
+        return res.status(FORBIDDEN).json({
+            error: forebiddenError,
         });
     }
 
@@ -182,9 +212,15 @@ router.put('/:id/toggleDuplicateHint', async (req: Request, res: Response) => {
 
 router.put('/:id/showHints', async (req: Request, res: Response) => {
     const game = await gameDao.getOne(req.params.id);
+    const userId = req.headers.authorization;
     if (!game) {
         return res.status(NOT_FOUND).json({
             error: gameNotFoundError,
+        });
+    }
+    if (game.roundHost !== userId) {
+        return res.status(FORBIDDEN).json({
+            error: forebiddenError,
         });
     }
 
@@ -200,6 +236,7 @@ router.put('/:id/showHints', async (req: Request, res: Response) => {
 
 router.put('/:id/guess', async (req: Request, res: Response) => {
     const { guess } = req.body;
+    const userId = req.headers.authorization;
     const game = await gameDao.getOne(req.params.id);
     if (!guess) {
         return res.status(BAD_REQUEST).json({
@@ -209,6 +246,11 @@ router.put('/:id/guess', async (req: Request, res: Response) => {
     if (!game) {
         return res.status(NOT_FOUND).json({
             error: gameNotFoundError,
+        });
+    }
+    if (game.currentGuesser !== userId) {
+        return res.status(FORBIDDEN).json({
+            error: forebiddenError,
         });
     }
 
@@ -224,10 +266,16 @@ router.put('/:id/guess', async (req: Request, res: Response) => {
 
 router.put('/:id/resolve', async (req: Request, res: Response) => {
     const { correct } = req.body;
+    const userId = req.headers.authorization;
     const game = await gameDao.getOne(req.params.id);
     if (!game) {
         return res.status(NOT_FOUND).json({
             error: gameNotFoundError,
+        });
+    }
+    if (game.players.findIndex(p => p.id === userId) === -1) {
+        return res.status(FORBIDDEN).json({
+            error: forebiddenError,
         });
     }
 
@@ -243,6 +291,18 @@ router.put('/:id/resolve', async (req: Request, res: Response) => {
 
 router.delete('/delete/:id', async (req: Request, res: Response) => {
     const { id } = req.params as ParamsDictionary;
+    const userId = req.headers.authorization;
+    const game = await gameDao.getOne(req.params.id);
+    if (!game) {
+        return res.status(NOT_FOUND).json({
+            error: gameNotFoundError,
+        });
+    }
+    if (game.host !== userId) {
+        return res.status(FORBIDDEN).json({
+            error: forebiddenError,
+        });
+    }
     await gameDao.delete(id);
     return res.status(OK).end();
 });
